+++ tests/test_measurements.py
@@ -526,7 +526,9 @@ def test_shots_single_measure_obs(shots, measure_f, obs, n_wires, mcmc, kernel_n
     """Tests that Lightning handles shots in a circuit where a single measurement of a common observable is performed at the end."""
 
     if (
-        shots is None or device_name in ("lightning.gpu", "lightning.kokkos", "lightning.tensor")
+        shots is None
+        or device_name
+        in ("lightning.gpu", "lightning.amdgpu", "lightning.kokkos", "lightning.tensor")
     ) and (mcmc or kernel_name != "Local"):
         pytest.skip(f"Device {device_name} does not have an mcmc option.")
 
@@ -539,7 +541,7 @@ def test_shots_single_measure_obs(shots, measure_f, obs, n_wires, mcmc, kernel_n
     if measure_f in (qml.expval, qml.var) and obs is None:
         pytest.skip("qml.expval, qml.var requires observable.")
 
-    if device_name in ("lightning.gpu", "lightning.kokkos"):
+    if device_name in ("lightning.gpu", "lightning.amdgpu", "lightning.kokkos"):
         dev = qml.device(device_name, wires=n_wires, seed=seed)
     elif device_name == "lightning.qubit":
         dev = qml.device(